feat: Refactor MoviePage component structure and update import paths
