+++ web/src/App.tsx
@@ -3,7 +3,7 @@ import {
   Navigate,
   RouterProvider,
 } from "react-router-dom";
-import MoviePage from "./pages/MoviePage/MoviePage";
+import MoviePage from "./pages/MoviePage";
 import ActorPage from "./pages/ActorListPage/ActorListPage";
 import HomePage from "./pages/HomePage/HomePage";
 import AppLayout from "./layouts/AppLayout";
